third_party/WebKit/LayoutTests/fullscreen/api/element-request-fullscreen-and-move-to-iframe.html

<!DOCTYPE html>
<title>Element#requestFullscreen() followed by moving the element into an iframe</title>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../trusted-click.js"></script>
<div id="log"></div>
<div id="target"></div>
<iframe allowfullscreen></iframe>
<script>
// Adapted from https://github.com/w3c/web-platform-tests/pull/4250
// TODO(foolip): Remove this test when the above is imported and passing.
async_test(t => {
  const target = document.getElementById("target");
  const iframeDoc = document.querySelector("iframe").contentDocument;

  iframeDoc.onfullscreenchange = t.unreached_func("fullscreenchange event in iframe");
  iframeDoc.onfullscreenerror = t.unreached_func("fullscreenerror event in iframe");
  document.onfullscreenchange = t.step_func_done(() => {
    assert_equals(document.fullscreenElement, null);
    assert_equals(iframeDoc.fullscreenElement, null);
  });
  document.onfullscreenerror = t.unreached_func("fullscreenerror event");

  trusted_click(t.step_func(() => {
    target.requestFullscreen();
    iframeDoc.body.appendChild(target);
  }), document.body);
});
</script>
